+++ src/lib.rs
@@ -11,50 +11,29 @@ pub use checks::available_check_names;
 
 #[allow(clippy::redundant_closure)]
 pub fn run(args: clap::ArgMatches, current_dir: &PathBuf) -> Result<Vec<Warning>, Box<dyn Error>> {
-    let mut dirs: Vec<PathBuf> = Vec::new();
     let mut file_paths: Vec<PathBuf> = Vec::new();
     let mut skip_checks: Vec<&str> = Vec::new();
+    let mut excluded_paths: Vec<PathBuf> = Vec::new();
+
+    let is_recursive = args.is_present("recursive");
 
     if let Some(skip) = args.values_of("skip") {
         skip_checks = skip.collect();
     }
 
-    if let Some(inputs) = args.values_of("input") {
-        dirs = inputs
-            .clone()
-            .filter_map(|s| fs_utils::canonicalize(s).ok())
-            .filter(|p| p.is_dir())
-            .collect();
-
-        file_paths = inputs
-            .filter_map(|s| fs_utils::canonicalize(s).ok())
-            .filter(|p| p.is_file())
-            .collect();
-    }
-
-    for dir_path in dirs {
-        let entries = dir_path.read_dir()?;
-
-        let mut dir_files: Vec<PathBuf> = entries
-            .filter_map(|e| e.ok())
-            .map(|e| e.path())
-            .filter(|path| FileEntry::is_env_file(path))
-            .collect();
-
-        file_paths.append(dir_files.as_mut());
-    }
-
-    // Removes files from paths if they should be excluded
     if let Some(excluded) = args.values_of("exclude") {
-        let excluded_paths: Vec<PathBuf> = excluded
+        excluded_paths = excluded
             .filter_map(|f| fs_utils::canonicalize(f).ok())
             .collect();
-
-        file_paths.retain(|path| !excluded_paths.contains(&path));
     }
 
-    file_paths.sort();
-    file_paths.dedup();
+    if let Some(inputs) = args.values_of("input") {
+        let input_paths = inputs
+            .filter_map(|s| fs_utils::canonicalize(s).ok())
+            .collect();
+
+        file_paths.extend(get_file_paths(input_paths, &excluded_paths, is_recursive));
+    }
 
     let mut warnings: Vec<Warning> = Vec::new();
 
@@ -76,6 +55,38 @@ pub fn run(args: clap::ArgMatches, current_dir: &PathBuf) -> Result<Vec<Warning>
     }
 
     Ok(warnings)
+}
+
+fn get_file_paths(
+    dir_entries: Vec<PathBuf>,
+    excludes: &[PathBuf],
+    is_recursive: bool,
+) -> Vec<PathBuf> {
+    let nested_paths: Vec<PathBuf> = dir_entries
+        .iter()
+        .filter(|entry| entry.is_dir())
+        .filter(|entry| !excludes.contains(entry))
+        .filter_map(|dir| dir.read_dir().ok())
+        .map(|read_dir| {
+            read_dir
+                .filter_map(|e| e.ok())
+                .map(|e| e.path())
+                .filter(|path| FileEntry::is_env_file(path) || (is_recursive && path.is_dir()))
+                .collect()
+        })
+        .flat_map(|dir_entries| get_file_paths(dir_entries, excludes, is_recursive))
+        .collect();
+
+    let mut file_paths: Vec<PathBuf> = dir_entries
+        .into_iter()
+        .filter(|entry| entry.is_file())
+        .filter(|entry| !excludes.contains(entry))
+        .collect();
+
+    file_paths.sort();
+    file_paths.extend(nested_paths);
+    file_paths.dedup();
+    file_paths
 }
 
 fn get_line_entries(fe: FileEntry, lines: Vec<String>) -> Vec<LineEntry> {