+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add build and test steps running on Windows [#216](https://github.com/dotenv-linter/dotenv-linter/pull/216) ([@mgrachev](https://github.com/mgrachev))
 
 ### 🔧 Changed
-- Optimize integration tests [#218](https://github.com/dotenv-linter/dotenv-linter/pull/218) ([@DDtKey](https://github.com/mgrachev))
+- Optimize integration tests [#218](https://github.com/dotenv-linter/dotenv-linter/pull/218) ([@mgrachev](https://github.com/mgrachev))
 
 ## [v2.0.0] - 2020-06-15
 ### 🚀 Added